+++ src/main/java/net/seabears/register/orders/Quantity.java
@@ -4,6 +4,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Quantity {
     @JsonProperty("item_id")
-    public long itemId;
+    public int itemId;
     public int amount;
 }
